pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuro-backend"
version = "0.1.0"
description = "Pomodoro timer API for the Nuro project."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{name = "Nuro"}]
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.13.2",
    "asyncpg==0.29.0",
    "psycopg[binary,pool]==3.2.1",
    "redis[hiredis]==5.0.7",
    "pydantic-settings==2.4.0",
    "pyjwt[crypto]==2.9.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "httpx==0.27.2",
    "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
    "black==24.8.0",
    "ruff==0.6.8",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.1.0",
    "fakeredis==2.23.4",
    "pytest-mock==3.14.0",
    "freezegun==1.5.1",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = "\\.pyi?$"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "ASYNC"]
ignore = ["B008"]
exclude = ["backend/alembic/versions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--strict-markers --disable-warnings"
testpaths = ["backend/tests"]

[tool.coverage.run]
source = ["backend/app"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
